Extract shared PyTorch layers from the SLi-Rec base into models/pytorch/layers.py
FcnNet, init_weight_ and the activation table are ports of base_model.py pieces
that every deeprec model needs, not sequential-specific ones. Moving them out of
sequential_base.py lets the upcoming xDeepFM port reuse them.
